main_ws/requirements.txt
google-genai
jinja2==3.1.6
numpy==1.26.4
pyyaml==6.0.3
qpsolvers==4.8.2
quadprog==0.1.13
roboticstoolbox-python @ git+https://github.com/Tsaimingchun14/robotics-toolbox-python.git@6601137adde87126a415e87755a263a567e6f7bf
rtb-data @ git+https://github.com/Tsaimingchun14/robotics-toolbox-python.git@6601137adde87126a415e87755a263a567e6f7bf#subdirectory=rtb-data
scipy==1.15.3
spatialgeometry==1.1.0
spatialmath-python==1.1.15
swift-sim==1.1.0
websockets==14.2
pyzmq
python-dotenv
empy==3.3.4
catkin-pkg==1.1.0
lark==1.3.1
